requirements.txt
Flask
openai
beautifulsoup4
lxml
requests
boto3
flask-bcrypt
flask-cors
scikit-learn
python-dotenv